Move model classes to their own files in order to comply with PSR-4
