+++ src/api/data/models/Alert.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class Alert
+{
+    public $header;
+
+    public $description;
+}
