+++ src/api/data/models/Connection.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class Connection
+{
+    public $departure;
+
+    public $arrival;
+
+    public $via;
+    // not compulsory
+    public $duration;
+}
