+++ src/api/data/models/DepartureArrival.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class DepartureArrival
+{
+    public $delay;
+
+    public $station;
+
+    public $time;
+
+    public $vehicle;
+
+    public $platform;
+
+    public $canceled;
+}
