+++ src/api/data/models/Station.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class Station
+{
+    public $_hafasId;
+
+    public $locationX;
+
+    public $locationY;
+
+    public $id;
+
+    public $name;
+}
