+++ src/api/data/models/Stop.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Irail\api\data\
+models;
+
+class Stop
+{
+    public $station;
+
+    public $time;
+
+    public $delay;
+
+    public $platform;
+
+    public $canceled;
+}
