+++ src/api/data/models/TrainCompositionInSegment.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class TrainCompositionInSegment
+{
+    public $origin;
+
+    public $destination;
+
+
+    /**
+     * @var $composition TrainComposition.
+     */
+    public $composition;
+}
