+++ src/api/data/models/TrainCompositionResult.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class TrainCompositionResult
+{
+    /**
+     * @var $segment TrainCompositionInSegment[] A list of all segments with their own composition for this train ride.
+     */
+    public $segment;
+}
