+++ src/api/data/models/TrainCompositionUnit.php
@@ -1,183 +1,9 @@
 <?php
 
-namespace Irail\api\data;
-
-/** Copyright (C) 2011 by iRail vzw/asbl
- * This file contains classes used in API responses.
- *
- * @author pieterc
- */
-
-//class Liveboard {
-//  public $station;
-
-//public $departure;
-
-//public $arrival;
-
-// }
-
-class Disturbance
-{
-    public $title;
-    public $description;
-    // public $attachment; // Not compulsory, commented to ensure null values don't cause issues in the printer
-    public $link; // Not compulsory
-    public $type;
-    public $timestamp;
-}
-
-class Connection
-{
-    public $departure;
-
-    public $arrival;
-
-    public $via;
-    // not compulsory
-    public $duration;
-}
-
-class Station
-{
-    public $_hafasId;
-
-    public $locationX;
-
-    public $locationY;
-
-    public $id;
-
-    public $name;
-}
-
-class DepartureArrival
-{
-    public $delay;
-
-    public $station;
-
-    public $time;
-
-    public $vehicle;
-
-    public $platform;
-
-    public $canceled;
-}
-
-class Platform
-{
-    public $name;
-
-    public $normal;
-
-
-    /**
-     * Platform constructor.
-     * @param $name
-     * @param $normal
-     */
-    public function __construct($name = null, $normal = null)
-    {
-        $this->name = $name;
-
-        $this->normal = $normal;
-    }
-}
-
-class Via
-{
-    public $arrival;
-
-    public $departure;
-
-    public $timeBetween;
-
-    public $station;
-
-    public $vehicle;
-}
-
-class Vehicle
-{
-    public $locationX;
-
-    public $locationY;
-
-    public $name;
-
-    public $shortname;
-}
-
-class ViaDepartureArrival
-{
-    public $time;
-
-    public $platform;
-
-    public $isExtraStop;
-}
-
-class Stop
-{
-    public $station;
-
-    public $time;
-
-    public $delay;
-
-    public $platform;
-
-    public $canceled;
-}
-
-class Alert
-{
-    public $header;
-
-    public $description;
-}
-
-class TrainCompositionResult
-{
-    /**
-     * @var $segment TrainCompositionInSegment[] A list of all segments with their own composition for this train ride.
-     */
-    public $segment;
-}
-
-
-class TrainCompositionInSegment
-{
-    public $origin;
-
-    public $destination;
-
-
-    /**
-     * @var $composition TrainComposition.
-     */
-    public $composition;
-}
-
-class TrainComposition
-{
-    /**
-     * @var String internal source of this data, for example "Atlas".
-     */
-    public $source;
-
-
-    /**
-     * @var TrainCompositionUnit[] the units in this composition.
-     */
-    public $unit;
-}
+namespace Irail\api\data\models;
 
 class TrainCompositionUnit
 {
-
     //-- Computed fields
 
     /**
@@ -301,21 +127,4 @@ class TrainCompositionUnit
      */
     public $hasPriorityPlaces;
     //-- Other types can be included dynamically but aren't guaranteed to be present
-}
-
-class RollingMaterialType
-{
-    /**
-     * @var $parent_type string the parent type, such as I6, M5, HLE27, AM86 ...
-     */
-    public $parent_type;
-
-    /**
-     * @var $sub_type string the sub type, such as A, B, BD, BDx, C, ...
-     */
-    public $sub_type;
-    /**
-     * @var $orientation string The orientation of the vehicle, LEFT (default) or RIGHT.
-     */
-    public $orientation;
 }
