+++ src/api/data/models/Vehicle.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class Vehicle
+{
+    public $locationX;
+
+    public $locationY;
+
+    public $name;
+
+    public $shortname;
+}
