+++ src/api/data/models/Via.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class Via
+{
+    public $arrival;
+
+    public $departure;
+
+    public $timeBetween;
+
+    public $station;
+
+    public $vehicle;
+}
