+++ src/api/data/models/ViaDepartureArrival.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Irail\api\data\models;
+
+class ViaDepartureArrival
+{
+    public $time;
+
+    public $platform;
+
+    public $isExtraStop;
+}
